backend/requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
langchain
langchain-community
langchain-text-splitters
sentence-transformers
chromadb
langchain-groq
langchain-huggingface
